change route and update component
